requirements.txt
python-dotenv
hupper
gradio
termcolor
openai
tenacity
supabase
langchain
langchain-community
langchain-openai
tiktoken
beautifulsoup4
faiss-cpu
pypdf
chardet
sentence-transformers